+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## Next release - 4.6
 
 - [#714](https://github.com/openshift/cluster-monitoring-operator/pull/714) Validate new/updated PrometheusRule custom resources against the prometheus-operator rule validation API.
+- [#799](https://github.com/openshift/cluster-monitoring-operator/pull/799) Rules federation support.